feat: global constellation globe background and transparent header on all pages
- Move globe from per-page (representatives) to global shell-level fixed background
- Globe renders at 20% opacity with bottom gradient fade on all pages except homepage
- Homepage keeps its own more prominent globe (25% authenticated, full anonymous)
- Header is now transparent on all pages, same scroll-based frosted glass behavior
- Globe position responds to sidebar collapse state (left offset transitions)
- Removes per-page globe code from CivicaDRepBrowse (handled globally now)
