+++ components/civica/CivicaShell.tsx
@@ -1,13 +1,19 @@
 'use client';
 
 import { Suspense, useEffect, useState, useCallback } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { usePathname, useRouter, useSearchParams } from 'next/navigation';
+import dynamic from 'next/dynamic';
 import { cn } from '@/lib/utils';
 import { SegmentProvider } from '@/components/providers/SegmentProvider';
 import { TierThemeProvider } from '@/components/providers/TierThemeProvider';
 import { CivicaHeader } from './CivicaHeader';
 import { CivicaBottomNav } from './CivicaBottomNav';
 import { CivicaSidebar } from './CivicaSidebar';
+
+const ConstellationScene = dynamic(
+  () => import('@/components/ConstellationScene').then((m) => ({ default: m.ConstellationScene })),
+  { ssr: false },
+);
 
 const SIDEBAR_STORAGE_KEY = 'governada_sidebar_collapsed';
 
@@ -49,6 +55,8 @@ function DeepLinkHandler() {
  * Sidebar is persona-adaptive via the nav config.
  */
 export function CivicaShell({ children }: { children: React.ReactNode }) {
+  const pathname = usePathname();
+  const isHome = pathname === '/';
   const [sidebarCollapsed, setSidebarCollapsed] = useState(false);
 
   useEffect(() => {
@@ -72,6 +80,24 @@ export function CivicaShell({ children }: { children: React.ReactNode }) {
         </Suspense>
         <CivicaHeader />
         <CivicaSidebar collapsed={sidebarCollapsed} onToggle={toggleSidebar} />
+
+        {/* Global constellation globe — subtle background on all pages except homepage */}
+        {!isHome && (
+          <div
+            className={cn(
+              'fixed inset-0 pointer-events-none z-0 transition-[left] duration-200',
+              sidebarCollapsed ? 'lg:left-16' : 'lg:left-60',
+            )}
+            aria-hidden="true"
+          >
+            <div className="absolute inset-0 opacity-20">
+              <ConstellationScene interactive={false} className="w-full h-full" />
+            </div>
+            {/* Gradient fade — globe is most visible at top, fades toward bottom */}
+            <div className="absolute inset-0 bg-gradient-to-b from-transparent via-transparent via-40% to-background" />
+          </div>
+        )}
+
         <main
           id="main-content"
           className={cn(
